Dev loading issues (#104)
* FIX split vue specific into another bundler, re-enable min-css loader to load avoid css injection

* FIX: imcomplete bundle rendering for vue pages
