+++ frontend/src/styles/custom.scss
@@ -7,6 +7,7 @@ $custom-green-light:  #e4f2e5;
 
 @font-face {
   font-family: 'GT Pressura';
+  font-display: 'swap';
   src:
     url('../fonts/GTPressura-Bold.woff2') format("woff2"),
     url('../fonts/GTPressura-Bold.woff') format("woff"),
