+++ frontend/webpack/webpack.config.dev.js
@@ -1,8 +1,8 @@
 const Path = require('path');
 const Webpack = require('webpack');
 const { merge } = require('webpack-merge');
-const StylelintPlugin = require('stylelint-webpack-plugin');
 const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const StylelintPlugin = require('stylelint-webpack-plugin');
 const ESLintPlugin = require('eslint-webpack-plugin');
 
 const common = require('./webpack.common.js');
@@ -33,10 +33,10 @@ module.exports = merge(common, {
       files: Path.join('src', '**/*.s?(a|c)ss'),
       fix: true,
     }),
-    new MiniCssExtractPlugin({filename: 'css/[name].css',}),
+    new MiniCssExtractPlugin({ filename: 'css/[name].css', }),
     new ESLintPlugin({
-        fix: true,  
-      }),
+      fix: true,
+    }),
   ],
   module: {
     rules: [
@@ -44,32 +44,7 @@ module.exports = merge(common, {
         test: /\.html$/i,
         loader: 'html-loader',
       },
-      {
-        test: /\.js$/,
-        include: Path.resolve(__dirname, '../src'),
-        loader: 'babel-loader',
-      },
-      {
-        test: /\.css$/,
-        loader: 'css-loader'
-      },
-      {
-        test: /.vue$/, 
-        loader: 'vue-loader'
-      },
-      {
-        test: /\.s(c|a)ss$/,
-        use: [
-          'vue-style-loader',
-          'css-loader',
-          {
-            loader: 'sass-loader',
-            options: {
-              implementation: require('sass'),
-            },
-          },
-        ],
-      },
+
     ],
-  },
+  }
 });
