+++ interlab/templates/vuejs.html
@@ -1,5 +1,9 @@
-{% extends 'base.html' %}
 
+{% extends 'base.html' %}
+{% load render_bundle from webpack_loader %}
+{% block header %}
+{% render_bundle 'vue' 'css' %}
+{% endblock header %}
 {% block content %}
 <div id="app">
     <v-app id="inspire">
@@ -28,3 +32,6 @@
     </v-app>
 </div>
 {% endblock content %}
+{% block js_footer %}
+{% render_bundle 'vue' 'js' %}
+{% endblock js_footer %}
